requirements.txt
streamlit
numpy
opencv-python-headless
scipy
Pillow
